company_chatbot/requirements.txt
langchain
langchain-community
langchain-huggingface
chromadb
fastapi
uvicorn
python-dotenv
streamlit
pypdf
torch
transformers
sentence-transformers